+++ apps/block_scout_web/test/block_scout_web/views/address_coin_balance_view_test.exs
@@ -20,13 +20,13 @@ defmodule BlockScoutWeb.AddressCoinBalanceViewTest do
 
   describe "delta_arrow/1" do
     test "return up pointing arrow for positive value" do
-      value = Decimal.new(123)
+      value = Wei.from(Decimal.new(123), :wei)
 
       assert AddressCoinBalanceView.delta_arrow(value) == "▲"
     end
 
     test "return down pointing arrow for negative value" do
-      value = Decimal.new(-123)
+      value = Wei.from(Decimal.new(-123), :wei)
 
       assert AddressCoinBalanceView.delta_arrow(value) == "▼"
     end
@@ -34,13 +34,13 @@ defmodule BlockScoutWeb.AddressCoinBalanceViewTest do
 
   describe "delta_sign/1" do
     test "return Positive for positive value" do
-      value = Decimal.new(123)
+      value = Wei.from(Decimal.new(123), :wei)
 
       assert AddressCoinBalanceView.delta_sign(value) == "Positive"
     end
 
     test "return Negative for negative value" do
-      value = Decimal.new(-123)
+      value = Wei.from(Decimal.new(-123), :wei)
 
       assert AddressCoinBalanceView.delta_sign(value) == "Negative"
     end
@@ -48,13 +48,13 @@ defmodule BlockScoutWeb.AddressCoinBalanceViewTest do
 
   describe "format_delta/1" do
     test "format positive values" do
-      value = Decimal.new(1_340_000_000_000_000_000)
+      value = Wei.from(Decimal.new(1_340_000_000_000_000_000), :wei)
 
       assert AddressCoinBalanceView.format_delta(value) == "1.34 POA"
     end
 
     test "format negative values" do
-      value = Decimal.new(-1_340_000_000_000_000_000)
+      value = Wei.from(Decimal.new(-1_340_000_000_000_000_000), :wei)
 
       assert AddressCoinBalanceView.format_delta(value) == "1.34 POA"
     end